update: Seperate The Dialog Files At Widgets Folder

+++ app/(plateform)/products/add-product/schema.ts
@@ -6,8 +6,8 @@ export const schema = z.object({
     barcode: z.string().optional(),
     unit: z.enum(["packet", "liter", "kilo"]),
     min: z.number().int().positive(),
-    storeCount: z.number().int().min(0),
-    marketCount: z.number().int().min(0),
+    storeCount: z.number().min(0),
+    marketCount: z.number().min(0),
     purchasePrice: z.number().min(0),
     sellingPrice: z.number().min(0),
 });
